refactor API: UptimeInfoAdapter accessible over UptimeInfo.h now (instead of Timer.h); remove STM32Cube impl. from DefaultUptimeInfoAdapter (reverted to former behavior for backward compatibility).
