+++ UptimeInfo.cpp
@@ -4,14 +4,12 @@
  *  Created on: 01.10.2013
  *      Author: niklausd
  */
-#include "Timer.h"
 #include "UptimeInfo.h"
 
 #ifdef ARDUINO
 #include "Arduino.h"
 #else
-//#include <sys/time.h>
-#include "main.h"
+#include <sys/time.h>
 #endif
 
 class DefaultUptimeInfoAdapter : public UptimeInfoAdapter
@@ -22,13 +20,12 @@ public:
 #ifdef ARDUINO
     return millis();
 #else
-///**
-// * @see http://stackoverflow.com/a/1952423
-// */
-//    struct timeval tp;
-//    gettimeofday(&tp, 0);
-//    unsigned long ms = tp.tv_sec * 1000 + tp.tv_usec / 1000;
-    unsigned long ms = HAL_GetTick();
+/**
+ * @see http://stackoverflow.com/a/1952423
+ */
+    struct timeval tp;
+    gettimeofday(&tp, 0);
+    unsigned long ms = tp.tv_sec * 1000 + tp.tv_usec / 1000;
     return ms;
 #endif
   }
